requirements/common.txt
evdev
python-uinput
click
pyyaml
